skills: drop SQLite ChartSkillInfo; merge into EzSkillStore.
Abandon unused analysis JSON path without import; single Realm skill facade for later Skillset-Cache writers.
