+++ osu.Game/EzOsuGame/Skills/DATA-FOLLOWUPS.md
@@ -1,15 +1,19 @@
 # Skillset / Dan DATA follow-ups
 
-Tracked so later PRs do not forget these. **EZ Realm is frozen at 9** for this track — fill existing columns/tables only.
+Tracked so later PRs stay light. **EZ Realm frozen at 9** for this track — fill existing columns/tables only.
 
 Do **not** re-encode these as `// TODO(data)` in product code. Open a dedicated PR/plan per row.
 
 | ID | Plan name | Scope | Status |
 |----|-----------|--------|--------|
-| — | SCHEMA EZ9 | Typed `EzBeatmapChartSkillInfo` + `EzPlayerDanSkillsetValue` in Realm | **Done in PR1** |
-| 1 | DATA-LeoBlack-ChartDan | Full LeoBlack chart-side aggregate dan (optional; not required for maintainability). Also fills cluster columns if still null. | Wish-list; may never ship |
-| 2 | DATA-Skillset-Cache | Writers for `EzPlayerDanSkillsetValue` so DualPanel / profile do not re-file every clear on read. Table already exists (EZ9). | Optional next |
+| — | SCHEMA EZ9 | Typed `EzBeatmapChartSkillInfo` + `EzPlayerDanSkillsetValue` in Realm | **Done** |
+| — | Debt: SQLite ChartSkillInfo | Abandoned analysis `chart_skill_info` JSON; no import/migrate; miss → recompute | **Done** |
+| — | Debt: single Realm skill facade | ChartSkillInfo CRUD merged into `EzSkillStore`; deleted `EzChartSkillInfoStore` | **Done** |
+| 1 | DATA-LeoBlack-ChartDan | Full LeoBlack chart-side aggregate dan (optional). Also fills cluster columns if still null. | Wish-list |
+| 2 | DATA-Skillset-Cache | Writers for existing `EzPlayerDanSkillsetValue` + Provider cache-hit (DualPanel / profile stop re-filing every clear on read) | **Next light PR** |
 | 3 | DATA-Dan-Headline-Anchor | 7K LN `anchoredSkillsetDans`-style side headline (only if product moves titles off side-level `GetDan`). | Optional |
 | 4 | DATA-ChartSkillInfo-Batch | Background / library-wide ChartSkillInfo warm (beyond on-demand DualPanel + clear hashes). | Optional |
+
+UI read path (all of HUD DualPanel/Radar, Analysis Wedge, LocalProfile Track, display tags): **`EzSkillProvider` only**.
 
 When starting any of the above, check this file and the matching Cursor/plan todos, then tick or remove the row here in the same PR.
