+++ osu.Game/EzOsuGame/Skills/EzSkillStore.cs
@@ -9,7 +9,8 @@ using osu.Game.Database;
 namespace osu.Game.EzOsuGame.Skills
 {
     /// <summary>
-    /// Read/write helpers for skill Realm objects. Does not open the game for the caller.
+    /// Single Realm persistence facade for mania skill metrics (MSD / SSR / Dan / ChartSkillInfo / reserved skillset cache).
+    /// Does not open the game for the caller. Evidence lists stay in local-profile SQLite.
     /// </summary>
     public sealed class EzSkillStore
     {
@@ -367,6 +368,83 @@ namespace osu.Game.EzOsuGame.Skills
             });
         }
 
+        /// <summary>
+        /// Typed Realm ChartSkillInfo (EZ≥9). Miss → callers recompute; no SQLite JSON fallback.
+        /// </summary>
+        public bool TryGetChartSkillInfo(string beatmapHash, out EzChartSkillInfo? info)
+        {
+            info = null;
+
+            if (string.IsNullOrEmpty(beatmapHash))
+                return false;
+
+            EzChartSkillInfo? fromRealm = null;
+
+            realmAccess.Run(r =>
+            {
+                var row = r.All<EzBeatmapChartSkillInfo>()
+                           .FirstOrDefault(v => v.BeatmapHash == beatmapHash
+                                                && v.InfoVersion == EzChartSkillInfo.VERSION);
+                if (row != null)
+                    fromRealm = chartSkillInfoToDto(row);
+            });
+
+            if (fromRealm == null)
+                return false;
+
+            info = fromRealm;
+            return true;
+        }
+
+        public void UpsertChartSkillInfo(string beatmapHash, EzChartSkillInfo info, Guid beatmapId = default, DateTimeOffset? computedAt = null)
+        {
+            ArgumentException.ThrowIfNullOrWhiteSpace(beatmapHash);
+            ArgumentNullException.ThrowIfNull(info);
+
+            DateTimeOffset at = computedAt ?? DateTimeOffset.UtcNow;
+
+            realmAccess.Write(r =>
+            {
+                var existing = r.All<EzBeatmapChartSkillInfo>()
+                                .Where(v => v.BeatmapHash == beatmapHash)
+                                .ToList();
+
+                foreach (var row in existing)
+                    r.Remove(row);
+
+                r.Add(new EzBeatmapChartSkillInfo
+                {
+                    BeatmapHash = beatmapHash,
+                    BeatmapId = beatmapId,
+                    InfoVersion = EzChartSkillInfo.VERSION,
+                    ComputedAt = at,
+                    PatternTagsJoined = joinPatterns(info.Patterns),
+                    JackDemand = info.JackDemand,
+                    JackShare = info.JackShare ?? -1,
+                    StreamShare = info.StreamShare ?? -1,
+                    TechCategory = info.TechCategory,
+                    ClusterTrill = info.ClusterTrill,
+                    HandstreamCluster = info.HandstreamCluster,
+                    HandstreamEndurance = info.HandstreamEndurance,
+                    TechScore = info.TechScore,
+                    ChordjackScore = info.ChordjackScore,
+                    JackScore = info.JackScore ?? -1,
+                    MotionRhythmBreak = info.Motion?.RhythmBreak ?? -1,
+                    MotionCrossHandTrill = info.Motion?.CrossHandTrill ?? -1,
+                    MotionMiniJack = info.Motion?.MiniJack ?? -1,
+                    MotionSameHand = info.Motion?.SameHand ?? -1,
+                    LnRatio = info.LnRatio ?? -1,
+                    Vibro = info.Vibro,
+                    DanEligible = info.DanEligible,
+                    LengthSeconds = info.LengthSeconds ?? -1,
+                    KeyCount = info.KeyCount ?? -1,
+                });
+            });
+        }
+
+        /// <summary>
+        /// Reserved for DATA-Skillset-Cache: read cached skillset tiles. Product UI still computes via Provider until that PR wires writers.
+        /// </summary>
         public IReadOnlyList<EzPlayerDanSkillsetValue> GetDanSkillsetValues(
             string username,
             int keyCount,
@@ -389,7 +467,7 @@ namespace osu.Game.EzOsuGame.Skills
         }
 
         /// <summary>
-        /// Replaces all skillset cache rows for one username/key/side bucket.
+        /// Reserved for DATA-Skillset-Cache: replace skillset cache rows for one username/key/side bucket.
         /// </summary>
         public void WriteDanSkillsetValues(string username, int keyCount, string side, IEnumerable<EzPlayerDanSkillsetValue> values)
         {
@@ -427,6 +505,67 @@ namespace osu.Game.EzOsuGame.Skills
                     });
                 }
             });
+        }
+
+        private const char pattern_separator = '\u001f';
+
+        private static string joinPatterns(string[]? patterns)
+        {
+            if (patterns == null || patterns.Length == 0)
+                return string.Empty;
+
+            return string.Join(pattern_separator, patterns.Where(static t => !string.IsNullOrWhiteSpace(t)));
+        }
+
+        private static string[] splitPatterns(string? joined)
+        {
+            if (string.IsNullOrEmpty(joined))
+                return [];
+
+            return joined.Split(pattern_separator, StringSplitOptions.RemoveEmptyEntries);
+        }
+
+        private static double? optionalDouble(double value)
+            => value < 0 ? null : value;
+
+        private static EzChartSkillInfo chartSkillInfoToDto(EzBeatmapChartSkillInfo row)
+        {
+            EzMotionFeatures? motion = null;
+
+            if (row.MotionRhythmBreak >= 0
+                || row.MotionCrossHandTrill >= 0
+                || row.MotionMiniJack >= 0
+                || row.MotionSameHand >= 0)
+            {
+                motion = new EzMotionFeatures
+                {
+                    RhythmBreak = row.MotionRhythmBreak < 0 ? 0 : row.MotionRhythmBreak,
+                    CrossHandTrill = row.MotionCrossHandTrill < 0 ? 0 : row.MotionCrossHandTrill,
+                    MiniJack = row.MotionMiniJack < 0 ? 0 : row.MotionMiniJack,
+                    SameHand = row.MotionSameHand < 0 ? 0 : row.MotionSameHand,
+                };
+            }
+
+            return new EzChartSkillInfo
+            {
+                Patterns = splitPatterns(row.PatternTagsJoined),
+                JackDemand = row.JackDemand,
+                JackShare = optionalDouble(row.JackShare),
+                StreamShare = optionalDouble(row.StreamShare),
+                TechCategory = row.TechCategory,
+                ClusterTrill = row.ClusterTrill,
+                HandstreamCluster = row.HandstreamCluster,
+                HandstreamEndurance = row.HandstreamEndurance,
+                TechScore = row.TechScore,
+                ChordjackScore = row.ChordjackScore,
+                JackScore = optionalDouble(row.JackScore),
+                Motion = motion,
+                LnRatio = optionalDouble(row.LnRatio),
+                Vibro = row.Vibro,
+                DanEligible = row.DanEligible,
+                LengthSeconds = optionalDouble(row.LengthSeconds),
+                KeyCount = row.KeyCount < 0 ? null : row.KeyCount,
+            };
         }
     }
 }
